Update fields to bigint for postgresql

+++ db/migrate/20160303013811_add_redirect_count_to_link.rb
@@ -1,5 +1,5 @@
 class AddRedirectCountToLink < ActiveRecord::Migration
   def change
-    add_column :links, :redirect_count, :integer, default: 0
+    add_column :links, :redirect_count, :bigint, default: 0
   end
 end
