+++ db/migrate/20160303014403_add_redirect_limit_to_links.rb
@@ -1,5 +1,5 @@
 class AddRedirectLimitToLinks < ActiveRecord::Migration
   def change
-    add_column :links, :redirect_limit, :integer, default: 0
+    add_column :links, :redirect_limit, :bigint, default: 0
   end
 end
