+++ db/migrate/20160303023752_add_expiration_time_to_links.rb
@@ -1,5 +1,5 @@
 class AddExpirationTimeToLinks < ActiveRecord::Migration
   def change
-    add_column :links, :expiration_time, :integer
+    add_column :links, :expiration_time, :bigint
   end
 end
